style: 更改

+++ src/main/java/icu/xiaohu/diet_recommend/recommend/core/ItemCF.java
@@ -13,18 +13,18 @@ import java.util.stream.Collectors;
 public class ItemCF {
 
     /**
-     * 方法描述: 推荐电影id列表
+     * 方法描述: 推荐饮食id列表
      *
-     * @param itemId 当前电影id
-     * @param list 用户电影评分数据
-     * @return {@link List<Integer>}
+     * @param itemId 当前饮食项id
+     * @param list 用户饮食评分数据
+     * @return {@link List<Long>}
      * @date 2023年02月02日 14:51:42
      */
-    public static List<Integer> recommend(Integer itemId, List<RelateDTO> list) {
+    public static List<Long> recommend(Long itemId, List<RelateDTO> list) {
         //按物品分组
-        Map<Integer, List<RelateDTO>>  itemMap=list.stream().collect(Collectors.groupingBy(RelateDTO::getItemId));
+        Map<Long, List<RelateDTO>>  itemMap = list.stream().collect(Collectors.groupingBy(RelateDTO::getItemId));
         //获取其他物品与当前物品的关系值
-        Map<Integer,Double>  itemDisMap = CoreMath.computeNeighbor(itemId, itemMap,1);
+        Map<Long,Double>  itemDisMap = CoreMath.computeNeighbor(itemId, itemMap,1);
         //获取关系最近物品
         double maxValue=Collections.max(itemDisMap.values());
         return itemDisMap.entrySet().stream().filter(e->e.getValue()==maxValue).map(Map.Entry::getKey).collect(Collectors.toList());
